Modularized DrugPopup Component
Created a new DrugPopup component for the items to use in DayViewScreen, in a folder with the component definition and the styles sheet file. Ran through yarn linting to get rid of any unnecessary imports and lines of code.
